+++ tutor/opts.py
@@ -39,15 +39,4 @@ class YamlParamType(click.ParamType):
             k, v = value.split("=")
         except ValueError:
             self.fail("'{}' is not of the form 'key=value'.".format(value), param, ctx)
-        return k, serialize.parse_value(v)
-
-
-key_value = click.option(
-    "-s",
-    "--set",
-    "set_",
-    type=YamlParamType(),
-    multiple=True,
-    metavar="KEY=VAL",
-    help="Set a configuration value (can be used multiple times)",
-)
+        return k, serialize.parse(v)